fix delete file function

+++ src/app/src/services/canvas.js
@@ -174,14 +174,16 @@ class CanvasService {
             cursorStyle: "pointer", // Cursor changes to indicate interaction
             mouseDownHandler: (eventData, transform, x, y) => {
                 const target = transform.target;
-                const url = new URL(window.location.href);
-                const project_id = url.searchParams.get('project_id');
+                if (target.layerType === 'image') {
+                    const url = new URL(window.location.href);
+                    const project_id = url.searchParams.get('project_id');
 
-                let data = {
-                    project_id,
-                    file_name: transform.target.src,
+                    let data = {
+                        project_id,
+                        file_name: target.getSrc(),
+                    }
+                    MediaService.deleteFile(data);
                 }
-                MediaService.deleteFile(data);
 
                 if (target) {
                     canvas.remove(target); // Remove the object from the canvas
